requirements.txt
PyPDF2
fpdf
youtube_transcript_api
langchain
python-dotenv
langchain-community
pypdf
beautifulsoup4
bs4
arxiv
httpx==0.23.0
pymupdf
langchain-text-splitters
langchain-openai
streamlit_chat
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
pandas
openai
langchain-groq==0.2.1
langchain_core
uvicorn
validators==0.28.1
unstructured
huggingface_hub
groq
streamlit==1.31.0
streamlit-option-menu==0.3.12
pillow==10.2.0
pytesseract==0.3.10
python-docx==1.0.1
streamlit_option_menu